@@ -1,4 +1,6 @@
-apply plugin: 'com.android.library'
+plugins {
+    id 'com.android.library'
+}
 
 def artifact = "usbserial.jar"
 
@@ -7,45 +9,48 @@ android {
     group = 'com.felhr.usbserial'
     version = project.VERSION_NAME
 
-    compileSdkVersion Integer.parseInt(project.ANDROID_BUILD_SDK_VERSION)
-
-    compileOptions {
-        encoding "UTF-8"
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
+    namespace 'com.felhr.usbserial'
+    compileSdk Integer.parseInt(project.ANDROID_BUILD_SDK_VERSION)
 
     defaultConfig {
-        minSdkVersion Integer.parseInt(project.ANDROID_BUILD_MIN_SDK_VERSION)
+        minSdk Integer.parseInt(project.ANDROID_BUILD_MIN_SDK_VERSION)
         targetSdkVersion Integer.parseInt(project.ANDROID_BUILD_TARGET_SDK_VERSION)
         archivesBaseName = "${project.name}-${project.VERSION_NAME}"
-        consumerProguardFiles 'proguard-rules.pro'
+
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
+        consumerProguardFiles "consumer-rules.pro"
     }
 
-    //noinspection GroovyAssignabilityCheck
     task deleteJar(type: Delete) {
-        delete "eclipse_lib/${artifact}"
+        delete "libs/${artifact}"
     }
 
     //noinspection GroovyAssignabilityCheck
     task createJar(type: org.gradle.api.tasks.Copy) {
         from('build/intermediates/bundles/release/')
-        into('eclipse_lib/')
+        into('libs/')
         include('classes.jar')
         rename('classes.jar', artifact)
     }
 
-    dependencies {
-        implementation 'com.annimon:stream:1.2.1'
-        implementation 'com.squareup.okio:okio:2.1.0'
-
-        androidTestImplementation 'com.android.support.test:runner:1.0.2'
-        androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-        androidTestImplementation 'junit:junit:4.12'
-
-        testImplementation 'junit:junit:4.12'
-
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
+}
 
-    createJar.dependsOn(deleteJar, build)
+dependencies {
+    implementation 'com.annimon:stream:1.2.2'
+    implementation 'com.squareup.okio:okio:3.3.0'
+
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
 }
